DDP-5758 - refactor utility classes: convert it to factories, so it is possible to override it
